deletion done

+++ app/(tabs)/index.tsx
@@ -58,6 +58,7 @@ export default function Index() {
   const incrementTracker = useTrackerStore(state => state.incrementTracker);
   const moveSectionBy = useSectionStore(state => state.moveSectionBy);
   const deleteSection = useSectionStore(state => state.deleteSection)
+  const removeTrackerFromSection = useSectionStore(state => state.removeTrackerFromSection)
   /* States */
   //modal states
   const [sectionModalOpen, setSectionModalOpen] = useState(false);
@@ -99,6 +100,8 @@ export default function Index() {
     return thresholdsToReturn;
   };
 
+  //Edit mode states
+  const [sectionToDelete, setSectionToDelete] = useState<string | null>(null); //warning state for deletion
   //Edit mode refs
   const panRefs = useRef<{ [key: string]: Animated.ValueXY }>({}); //ref to all sections
   const pan = useRef(new Animated.ValueXY()).current; //ref to section being moved
@@ -193,7 +196,7 @@ export default function Index() {
   //function to get height and pos of section given title ()
   const getSectInfo = (sectionTitle: string): {height: number, position: number} => { //ONLY CALL WHEN ALREADY UNWRAPPED SECTION
     const section : Section =  sections.find((s) => s.sectionTitle === sectionTitle && s.timePeriod === selected)!
-    var sectHeight : number = 51.666 + 10; //(30) {height in theory} + (10) {padding size} + 11.666 (unaccounted for in top)
+    var sectHeight : number = 56.666 + 10; //(35) {height in theory} + (10) {padding size} + 11.666 (unaccounted for in top)
     var position : number = -1;
     if(section){
       const rows = Math.ceil((section.trackers.length +1)/ 4);
@@ -650,7 +653,7 @@ export default function Index() {
                 pan.getLayout(), //stored in pan object created by useRef earlier
                 {borderWidth: 1,
                   borderRadius: 8,
-                  borderColor: editMode ? currentTheme["lowOpacityWhite"] : 'transparent',
+                  borderColor: editMode ? currentTheme["white"] : 'transparent',
                   marginTop: section.position === 0 ? 30 : 15, //num1 from circle, num2 from other sections
                   paddingVertical: 10,
                   width: '100%',
@@ -671,6 +674,7 @@ export default function Index() {
                   alignContent: 'center',
                   justifyContent: 'center',
                   flexDirection: 'row',
+                  marginBottom: 5
                 }
               ]}
               >
@@ -688,11 +692,28 @@ export default function Index() {
                   right: 10,
                 }
               ]}
-              onPress={() =>
-                deleteSection(section.sectionTitle,section.timePeriod)
+              onPress={() =>{
+                Alert.alert(
+                  "Delete Section?",
+                  section.sectionTitle + " will be deleted permanently",
+                  [
+                    {
+                      text: "Cancel",
+                      style: "cancel",
+                    },
+                    {
+                      text: "OK",
+                      onPress: () => {
+                        deleteSection(section.sectionTitle,section.timePeriod)
+                      },
+                    },
+                  ],
+                  { cancelable: true }
+                );
+              }
               }>
                 <Feather
-                name="minus-circle"
+                name="x-circle"
                 size={30}
                 color={currentTheme['white']}
                 />
@@ -702,17 +723,26 @@ export default function Index() {
               {/* Section's Row of Tracker Icons */}
               <View style={styles.iconRow}>
                 {section.trackers.map((tracker) => (
+                  <View
+                    key = {tracker.trackerName + tracker.timePeriod}
+                    style = {[
+                      {flexDirection: 'row'}
+                    ]}
+                  >
+                    
                     
                     <Pressable
-                    key={tracker.trackerName + tracker.timePeriod}
                         // Single tap increment 
                         onPress={() => {
                             if (!editMode) { // don’t increment while you’re dragging sections
                             incrementTracker(tracker.trackerName, tracker.timePeriod);
+                            }else{
+
                             }
                         }}
                         // Hold press opens edit tracker
                         onLongPress={() => {
+                          if(!editMode){
                             router.push({
                             pathname: "/editTracker",
                             params: {
@@ -722,10 +752,12 @@ export default function Index() {
                                 image:    getIconInfo(tracker.icon).name,
                             },
                             });
+                          }
                         }}
                     style={[
                       squareIconButtonStyle(itemSize),
                       {
+                        borderColor: editMode ? currentTheme.white : currentTheme.dimgray,
                         backgroundColor: hexToRgba( 
                             // Set to 0 for transparency                          
                           getIconInfo(tracker.icon).color, 0               
@@ -755,8 +787,33 @@ export default function Index() {
                   
                     {getImage(tracker, 40).icon}
                   </Pressable>
+                  {/* Cross */}
+                  {editMode &&(
+                    <TouchableOpacity
+                    onPress={() => {
+                      removeTrackerFromSection(section.sectionTitle,selected,tracker)
+                    }}
+                    style = {[
+                      {position: 'absolute',
+                        backgroundColor: currentTheme['101010'],
+                        right: -4,
+                        top: 0,
+                        borderRadius: spacing,
+                      }
+                    ]}> 
+                      <Feather
+                          name="x-circle"
+                          size={spacing*2}
+                          color={currentTheme['white']}
+                          style={{ margin: -1 }} 
+                        />
+                    </TouchableOpacity>
+                  )}
+                  
+                  </View>
                   
                 ))}
+
 
                 {/* Plus button to open modal and store section */}
                 <Pressable
@@ -770,9 +827,10 @@ export default function Index() {
                 >
                   <AntDesign name="plus" size={30} color={currentTheme.white} />
                 </Pressable>
-              </View>
+                </View>
             </Animated.View>
             </View>
+          
             )
             })}
         
@@ -819,7 +877,6 @@ export default function Index() {
                 {trackers
                   .filter((tracker) => tracker.timePeriod === selected) // Filter trackers by selected time period
                   .map((tracker) => (
-                    
                   <TouchableOpacity
                     key={tracker.trackerName + tracker.timePeriod}
                     onPress={() => {
